Remove shebang from load-omnibus-toolchain.sh
The shebang in this script makes the environment variables only
accessible to the shell which it invokes and any subshells. This can be
an acceptible behavior provided the script is used to execute any
commands which need to be contained within the context of this
particular subshell, however parent shells will not inherit the
environment variables even when they are exported.